line 19: remove 'const' before MaterialApp
I was following the tutorials but I get the error " you can not invoke const..." I checked thoroughly to confirm my code is exactly like this one but the error couldnt go and the app failed to load. I checked stack overflow and I luckily found out that all I have to do is remove the const. I know there could be a better and more technical way to put this out, but I'm only a beginner. Hoping for your response, thank you for the lessons