Add new `startup_namer` codelab (#1234)
This is the code for an upcoming refresh of the 'Your first Flutter app' codelab, also known as `startup_namer`.

The two codelabs (the old `startup_namer` and the new `namer`) will probably need to coexist for a while — but not too long. The new codelab will be used both on `codelabs.developers.google.com` and on `flutter.dev`.