Fixing mislabelled file in step 5
In Step 5 of the code lab, `main.dart` is referenced by name, despite the step requiring a change to `home_page.dart` instead. I found this a bit confusing when I was looking for the `HomePage` method in `main.dart`, so I think it makes sense to correct this as proposed.