`in_app_purchases` codelab update (#2295)
This PR is for a matching docs update. There are only minor changes in
the code:
1. `googlePlayProjectName` -> `googleCloudProjectId` const rename to
make it more appropriate. It is used for Pub/Sub to which Google Play
publishes purchase-related events but it's on Google Cloud so we need
the ID of the GCP project.
2. Steps `09` and `10` were misaligned relative to the instructions and
should be steps `10` and `11` instead.
3. Minor update to backend readme. The `constants.dart` already exists,
there was a typo in one of the placeholder values, and `constants.dart`
uses single quotes.

## Pre-launch Checklist

- [x] I read the [Effective Dart: Style] _recently_, and have followed
its advice.
- [x] I signed the [CLA].
- [x] I updated/added relevant documentation (doc comments with `///`).
- [x] All existing and new tests are passing.

If you need help, consider asking for advice on the #hackers-devrel
channel on [Discord].

<!-- Links -->
[Effective Dart: Style]:
https://dart.dev/guides/language/effective-dart/style
[CLA]: https://cla.developers.google.com/
[Discord]:
https://github.com/flutter/flutter/blob/master/docs/contributing/Chat.md