Rebuild script for `tfserving-flutter` (#2307)
This adds a rebuild script for `tfserving-flutter`. In the process it
extends the `codelab_rebuild` tooling to support running `protoc` to
generate dart code from protobuf definitions. Oh, and upgrading the
codelab to latest Dart and protobuf libraries.

## Pre-launch Checklist

- [x] I read the [Effective Dart: Style] _recently_, and have followed
its advice.
- [x] I signed the [CLA].
- [x] I updated/added relevant documentation (doc comments with `///`).
- [x] All existing and new tests are passing.

If you need help, consider asking for advice on the #hackers-devrel
channel on [Discord].

<!-- Links -->
[Effective Dart: Style]:
https://dart.dev/guides/language/effective-dart/style
[CLA]: https://cla.developers.google.com/
[Discord]:
https://github.com/flutter/flutter/blob/master/docs/contributing/Chat.md